requirements.txt
fastapi
uvicorn
pydantic
Pillow
opencv-python
torch
torchvision
numpy
matplotlib
scikit-image
segment-anything